+++ core/teams/snapshot.go
@@ -120,6 +120,12 @@ func (e *SnapshottedEngine) serialiseTeams() ([]byte, error) {
 			CreatedAt: team.CreatedAt.UnixNano(),
 			Closed:    team.Closed,
 		}
+
+		teamSnapshot.AllowList = make([]string, 0, len(team.AllowList))
+		for _, partyID := range team.AllowList {
+			teamSnapshot.AllowList = append(teamSnapshot.AllowList, partyID.String())
+		}
+
 		teamsSnapshot = append(teamsSnapshot, teamSnapshot)
 	}